j. Lab10/requirements.txt
asgiref==3.6.0
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
daphne==4.0.0
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.5
django-cleanup==7.0.0
django-cors-headers==3.14.0
django-environ==0.10.0
django-filter==22.1
django-mixins==0.0.10
django-phonenumber-field==7.0.2
django-redis==5.2.0
django-templated-mail==1.1.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.5
filelock==3.10.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
msgpack==1.0.5
oauthlib==3.2.2
packaging==23.0
phonenumbers==8.13.7
Pillow==9.4.0
platformdirs==3.1.1
psycopg2==2.9.5
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.1.1
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
redis==4.5.3
repos==0.1.16
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.1
sqlparse==0.4.3
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
utils==1.0.1
uuid==1.30
virtualenv==20.21.0
zope.interface==6.0
